# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Next.js Go Monorepo Kit API
  version: 0.1.0
  description: |
    OpenAPI contract for the starter Go API used by the frontend app.
servers:
  - url: http://localhost:8080
    description: Local development
tags:
  - name: Health
  - name: Auth
  - name: Analytics
  - name: Resources
  - name: Entries
paths:
  /health:
    get:
      tags: [Health]
      operationId: getHealth
      summary: Get API health status
      responses:
        "200":
          description: Health status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /api/v1/auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Auth session payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/auth/forgot-password:
    post:
      tags: [Auth]
      operationId: forgotPassword
      summary: Request a password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "204":
          description: Password reset accepted
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      operationId: resetPassword
      summary: Reset a password with a token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "204":
          description: Password updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/auth/verify-email:
    post:
      tags: [Auth]
      operationId: verifyEmail
      summary: Verify an email address with a token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailInput"
      responses:
        "204":
          description: Email verified
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/auth/resend-verification:
    post:
      tags: [Auth]
      operationId: resendVerificationEmail
      summary: Resend email verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendVerificationInput"
      responses:
        "204":
          description: Verification email requested
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Get the currently authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/analytics/overview:
    get:
      tags: [Analytics]
      operationId: getAnalyticsOverview
      summary: Get owner dashboard analytics overview
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Analytics overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsOverviewResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "501":
          $ref: "#/components/responses/NotImplemented"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/v1/resources:
    get:
      tags: [Resources]
      operationId: listResources
      summary: List public resources
      responses:
        "200":
          description: Resource list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceListResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/resources/{slug}:
    parameters:
      - $ref: "#/components/parameters/Slug"
    get:
      tags: [Resources]
      operationId: getResource
      summary: Get a resource by slug
      responses:
        "200":
          description: Resource detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/resources/{slug}/entries:
    parameters:
      - $ref: "#/components/parameters/Slug"
    get:
      tags: [Entries]
      operationId: listEntriesByResource
      summary: List public entries for a resource
      responses:
        "200":
          description: Entry list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryListResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Slug:
      name: slug
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication failed or is missing
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict with existing data
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotImplemented:
      description: Feature not configured
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadGateway:
      description: Upstream integration failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
    HealthStatus:
      type: object
      additionalProperties: false
      required: [name, env, version, database, timestamp]
      properties:
        name:
          type: string
        env:
          type: string
        version:
          type: string
        database:
          type: string
        timestamp:
          type: string
          format: date-time
    Resource:
      type: object
      additionalProperties: false
      required:
        - id
        - owner_id
        - slug
        - title
        - description
        - visibility
        - status
        - locale
        - entry_count
        - estimated_minutes
        - created_at
        - updated_at
      properties:
        id:
          type: string
        owner_id:
          type: string
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        visibility:
          type: string
          enum: [private, unlisted, public]
        status:
          type: string
          enum: [draft, published, archived]
        locale:
          type: string
        entry_count:
          type: integer
        estimated_minutes:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Entry:
      type: object
      additionalProperties: false
      required:
        - id
        - resource_id
        - position
        - title
        - content
        - created_at
        - updated_at
      properties:
        id:
          type: string
        resource_id:
          type: string
        position:
          type: integer
        title:
          type: string
        content:
          type: string
        details:
          type: string
        notes:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    User:
      type: object
      additionalProperties: false
      required:
        - id
        - username
        - email
        - role
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          type: string
        status:
          type: string
        last_login_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AnalyticsOverview:
      type: object
      additionalProperties: false
      required:
        - visitors
        - pageviews
        - views_per_visit
        - bounce_rate
        - visit_duration
        - date_range
        - source
      properties:
        visitors:
          type: integer
        pageviews:
          type: integer
        views_per_visit:
          type: string
        bounce_rate:
          type: string
        visit_duration:
          type: string
        date_range:
          type: string
        source:
          type: string
    RegisterInput:
      type: object
      additionalProperties: false
      required: [username, email, password]
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        full_name:
          type: string
    LoginInput:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    ForgotPasswordInput:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordInput:
      type: object
      additionalProperties: false
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
    VerifyEmailInput:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
    ResendVerificationInput:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          format: email
    UserResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/User"
    ResourceResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Resource"
    ResourceListResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Resource"
    EntryListResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Entry"
    AnalyticsOverviewResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AnalyticsOverview"
    AuthResponse:
      type: object
      additionalProperties: false
      required: [access_token, token_type, expires_at, user]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          default: Bearer
        expires_at:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
    RegisterResponse:
      type: object
      additionalProperties: false
      required: [message, user]
      properties:
        message:
          type: string
        user:
          $ref: "#/components/schemas/User"